requirements.txt
pyTelegramBotAPI
requests
Pillow
moviepy
e621py